Renamed navbar to Navbar
