+++ apps/web/src/components/Navbar.tsx
