+++ apps/web/src/components/index.ts
@@ -5,4 +5,4 @@ export { MakeFirstUpper } from './MakeFirstUpper';
 export { MakeLastUpper } from './MakeLastUpper';
 export { LetterSpacing } from './LetterSpacing';
 export { FontDropdown } from './FontDropdown';
-export { Navbar } from './navbar';
+export { Navbar } from './Navbar';
